Moved SegmentedPeristentQueue's SerDes in its own class
